requirements.txt
pytest
flake8
mypy
psycopg2-binary
